WIP: need to handle font in dark mode
